+++ generators/typescript/sdk/generator/src/SdkGenerator.ts
@@ -2,6 +2,7 @@ import { AbsoluteFilePath } from "@fern-api/fs-utils";
 import { FernGeneratorExec } from "@fern-fern/generator-exec-sdk";
 import * as FernGeneratorExecSerializers from "@fern-fern/generator-exec-sdk/serialization";
 import { ExampleEndpointCall, HttpEndpoint, HttpService, IntermediateRepresentation } from "@fern-fern/ir-sdk/api";
+import { FdrSnippetTemplate, FdrSnippetTemplateClient, FdrSnippetTemplateEnvironment } from "@fern-fern/snippet-sdk";
 import {
     BundledTypescriptProject,
     convertExportedFilePathToFilePath,
@@ -35,6 +36,7 @@ import { TypeSchemaGenerator } from "@fern-typescript/type-schema-generator";
 import { writeFile } from "fs/promises";
 import { Directory, Project, SourceFile, ts } from "ts-morph";
 import urlJoin from "url-join";
+import { v4 as uuidv4 } from "uuid";
 import { SdkContextImpl } from "./contexts/SdkContextImpl";
 import { EndpointDeclarationReferencer } from "./declaration-referencers/EndpointDeclarationReferencer";
 import { EnvironmentsDeclarationReferencer } from "./declaration-referencers/EnvironmentsDeclarationReferencer";
@@ -46,6 +48,7 @@ import { SdkInlinedRequestBodyDeclarationReferencer } from "./declaration-refere
 import { TimeoutSdkErrorDeclarationReferencer } from "./declaration-referencers/TimeoutSdkErrorDeclarationReferencer";
 import { TypeDeclarationReferencer } from "./declaration-referencers/TypeDeclarationReferencer";
 import { ReferenceGenerator, ReferenceParameterDeclaration } from "./ReferenceGenerator";
+import { TemplateGenerator } from "./TemplateGenerator";
 import { JestTestGenerator } from "./test-generator/JestTestGenerator";
 
 const FILE_HEADER = `/**
@@ -71,6 +74,7 @@ export declare namespace SdkGenerator {
     export interface Config {
         whitelabel: boolean;
         snippetFilepath: AbsoluteFilePath | undefined;
+        snippetTemplateFilepath: AbsoluteFilePath | undefined;
         shouldBundle: boolean;
         shouldUseBrandedStringAliases: boolean;
         isPackagePrivate: boolean;
@@ -95,6 +99,9 @@ export declare namespace SdkGenerator {
         retainOriginalCasing: boolean;
         allowExtraFields: boolean;
         writeUnitTests: boolean;
+        executionEnvironment: "local" | "dev" | "prod";
+        organization: string;
+        apiName: string;
     }
 }
 
@@ -109,6 +116,7 @@ export class SdkGenerator {
     private extraScripts: Record<string, string> = {};
 
     private endpointSnippets: FernGeneratorExec.Endpoint[] = [];
+    private endpointSnippetTemplates: FdrSnippetTemplate.SnippetRegistryEntry[] = [];
 
     private project: Project;
     private rootDirectory: Directory;
@@ -146,6 +154,7 @@ export class SdkGenerator {
     private genericAPISdkErrorGenerator: GenericAPISdkErrorGenerator;
     private timeoutSdkErrorGenerator: TimeoutSdkErrorGenerator;
     private jestTestGenerator: JestTestGenerator;
+    private FdrClient: FdrSnippetTemplateClient | undefined;
 
     constructor({
         namespaceExport,
@@ -305,6 +314,16 @@ export class SdkGenerator {
             this.rootDirectory,
             this.config.writeUnitTests
         );
+
+        this.FdrClient =
+            this.config.executionEnvironment !== "local"
+                ? new FdrSnippetTemplateClient({
+                      environment:
+                          this.config.executionEnvironment === "dev"
+                              ? FdrSnippetTemplateEnvironment.Dev
+                              : FdrSnippetTemplateEnvironment.Prod
+                  })
+                : undefined;
     }
 
     public async generate(): Promise<TypescriptProject> {
@@ -370,6 +389,20 @@ export class SdkGenerator {
                 this.config.snippetFilepath,
                 JSON.stringify(await FernGeneratorExecSerializers.Snippets.jsonOrThrow(snippets), undefined, 4)
             );
+            if (this.config.snippetTemplateFilepath != null) {
+                await writeFile(
+                    this.config.snippetTemplateFilepath,
+                    JSON.stringify(this.endpointSnippetTemplates, undefined, 4)
+                );
+                if (this.FdrClient != null) {
+                    await this.FdrClient.templates.registerBatch({
+                        orgId: this.config.organization,
+                        apiId: this.config.apiName,
+                        apiDefinitionId: uuidv4(),
+                        snippets: this.endpointSnippetTemplates
+                    });
+                }
+            }
             if (this.config.includeApiReference && refGenerator !== undefined) {
                 this.extraFiles["reference.md"] = refGenerator.write();
             }
@@ -641,6 +674,49 @@ export class SdkGenerator {
                         endpoint.displayName ?? endpoint.name.pascalCase.unsafeName
                     );
                 }
+
+                if (this.config.snippetTemplateFilepath != null && this.npmPackage != null) {
+                    const project = new Project({
+                        useInMemoryFileSystem: true
+                    });
+                    const sourceFile = project.createSourceFile("snippet-test", undefined, { overwrite: true });
+                    const importsManager = new ImportsManager();
+                    const endpointContext = this.generateSdkContext(
+                        { sourceFile, importsManager },
+                        { isForSnippet: true }
+                    );
+
+                    const clientSourceFile = project.createSourceFile("snippet-client", undefined, { overwrite: true });
+                    const clientImportsManager = new ImportsManager();
+                    const clientContext = this.generateSdkContext(
+                        { sourceFile: clientSourceFile, importsManager: clientImportsManager },
+                        { isForSnippet: true }
+                    );
+
+                    const snippetTemplate = new TemplateGenerator({
+                        endpointContext,
+                        clientContext,
+                        npmPackage: this.npmPackage,
+                        endpoint,
+                        packageId,
+                        rootPackageId: rootPackage,
+                        retainOriginalCasing: this.config.retainOriginalCasing
+                    }).generateSnippetTemplate();
+                    if (snippetTemplate != null) {
+                        this.endpointSnippetTemplates.push({
+                            sdk: FdrSnippetTemplate.Sdk.typescript({
+                                package: this.npmPackage.packageName,
+                                version: this.npmPackage.version
+                            }),
+                            endpointId: {
+                                path: FernGeneratorExec.EndpointPath(this.getFullPathForEndpoint(endpoint)),
+                                method: endpoint.method
+                            },
+                            snippetTemplate
+                        });
+                    }
+                }
+
                 const example = endpoint.examples[0];
                 if (example != null) {
                     const snippet = this.withSnippet({
@@ -721,7 +797,7 @@ export class SdkGenerator {
                             // clientPath: getTextOfTsNode(statement),
                             clientPath: statement,
                             functionPath: serviceFilepath,
-                            functionName: endpoint.name.camelCase.unsafeName,
+                            functionName: this.getEndpointFunctionName(endpoint),
                             returnType,
                             parameters,
                             codeSnippet: referenceSnippet,
@@ -731,6 +807,10 @@ export class SdkGenerator {
                 }
             }
         });
+    }
+
+    private getEndpointFunctionName(endpoint: HttpEndpoint): string {
+        return endpoint.name.camelCase.unsafeName;
     }
 
     // TODO(dsinghvi): HACKHACK Move this to IR