+++ generators/typescript/sdk/generator/src/__test__/__mocks__/Mocks.ts
@@ -30,8 +30,12 @@ export const Mocks = {
     },
     sdkConfig(partialConfig?: Partial<SdkGenerator.Config>): SdkGenerator.Config {
         return {
+            organization: "test",
+            executionEnvironment: "local",
+            apiName: "test",
             whitelabel: false,
             snippetFilepath: undefined,
+            snippetTemplateFilepath: undefined,
             shouldUseBrandedStringAliases: false,
             isPackagePrivate: false,
             neverThrowErrors: false,