+++ generators/typescript/sdk/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.16.0-rc0] - 2024-04-24
+
+- Feature: The generator now registers snippet templates which can be used for dynamic
+  SDK code snippet generation.
+
 ## [0.15.1-rc1] - 2024-04-24
 
 - Improvement: Earlier for inlined request exports, we were doing the following: